label-studio/ml-backend/requirements-simple.txt
# Alternative requirements without label-studio-ml-backend
# Core dependencies
numpy>=1.21.0,<1.25.0
opencv-python-headless>=4.5.0,<5.0.0
pillow>=9.0.0,<11.0.0
requests>=2.25.0,<3.0.0

# ONNX Runtime for inference
onnxruntime>=1.12.0,<2.0.0

# Flask web framework
flask>=2.0.0,<3.0.0
flask-cors>=3.0.0,<5.0.0

# RTMLib for pose estimation
rtmlib>=0.0.13

# Utilities
tqdm>=4.60.0,<5.0.0

# Development and testing (optional)
pytest>=7.0.0,<8.0.0
pytest-mock>=3.6.0,<4.0.0
